+++ Tests/View/JsonpHandlerTest.php
@@ -17,6 +17,7 @@ use FOS\RestBundle\View\ViewHandler;
 use PHPUnit\Framework\TestCase;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\RequestStack;
+use Symfony\Component\Templating\EngineInterface;
 
 /**
  * Jsonp handler test.
@@ -35,7 +36,7 @@ class JsonpHandlerTest extends TestCase
     {
         $this->router = $this->getMockBuilder('Symfony\Component\Routing\RouterInterface')->getMock();
         $this->serializer = $this->getMockBuilder('FOS\RestBundle\Serializer\Serializer')->getMock();
-        $this->templating = $this->getMockBuilder('Symfony\Bundle\FrameworkBundle\Templating\EngineInterface')->getMock();
+        $this->templating = $this->getMockBuilder(EngineInterface::class)->getMock();
         $this->requestStack = new RequestStack();
     }