+++ Tests/View/ViewHandlerTest.php
@@ -20,6 +20,7 @@ use Symfony\Component\Form\FormView;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\RequestStack;
 use Symfony\Component\HttpFoundation\Response;
+use Symfony\Component\Templating\EngineInterface;
 
 /**
  * View test.
@@ -37,7 +38,7 @@ class ViewHandlerTest extends TestCase
     {
         $this->router = $this->getMockBuilder('Symfony\Component\Routing\RouterInterface')->getMock();
         $this->serializer = $this->getMockBuilder('FOS\RestBundle\Serializer\Serializer')->getMock();
-        $this->templating = $this->getMockBuilder('Symfony\Bundle\FrameworkBundle\Templating\EngineInterface')->getMock();
+        $this->templating = $this->getMockBuilder(EngineInterface::class)->getMock();
         $this->requestStack = new RequestStack();
     }
 
@@ -566,7 +567,7 @@ class ViewHandlerTest extends TestCase
 
     /**
      * @expectedException \LogicException
-     * @expectedExceptionMessage An instance of Symfony\Bundle\FrameworkBundle\Templating\EngineInterface must be injected in FOS\RestBundle\View\ViewHandler to render templates.
+     * @expectedExceptionMessage An instance of Symfony\Component\Templating\EngineInterface must be injected in FOS\RestBundle\View\ViewHandler to render templates.
      */
     public function testTemplatingNotInjected()
     {